requirements.txt
aniso8601==9.0.1
attrs==22.2.0
autopep8==2.0.2
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
exceptiongroup==1.1.1
Flask==2.2.3
Flask-Admin==1.6.1
Flask-RESTful==0.3.9
Flask-SQLAlchemy==3.0.3
greenlet==2.0.2
idna==3.4
importlib-metadata==6.0.0
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
packaging==23.0
Pillow==9.4.0
pluggy==1.0.0
pycodestyle==2.10.0
pytest==7.2.2
pytz==2022.7.1
requests==2.28.2
six==1.16.0
SQLAlchemy==2.0.6
tomli==2.0.1
typing_extensions==4.5.0
urllib3==1.26.15
Werkzeug==2.2.3
WTForms==3.0.1
zipp==3.15.0
